fix: updated test case
